+++ src/features/home/approve.feature.ts
@@ -1,14 +1,14 @@
 import { approvePetition } from 'apis/petition.api';
 import { useMutation, useQueryClient } from 'react-query';
 import * as KEY from 'constants/key.constant';
-import { useSuccesToast } from 'hooks/useToast';
+import { toast } from 'react-toastify';
 
 export const ApproveFeature = (petitionId: number) => {
   const queryClient = useQueryClient();
 
   const { mutate } = useMutation(approvePetition, {
     onSuccess: () => {
-      useSuccesToast('동의 완료');
+      toast.success('동의 완료');
       queryClient.invalidateQueries([KEY.PETITION]);
     },
     onError: (err) => {
