+++ src/features/home/write.feature.ts
@@ -1,8 +1,8 @@
 import { writePetition } from 'apis/petition.api';
 import { useModal } from 'hooks/useModal';
-import { useSuccesToast } from 'hooks/useToast';
 import { useMutation } from 'react-query';
 import { useNavigate } from 'react-router-dom';
+import { toast } from 'react-toastify';
 import { WritePetitionType } from 'types/petition.type';
 
 export const WriteFeature = (petitionData: WritePetitionType) => {
@@ -10,7 +10,7 @@ export const WriteFeature = (petitionData: WritePetitionType) => {
   const { closeModal } = useModal();
   const { mutate } = useMutation(writePetition, {
     onSuccess: () => {
-      useSuccesToast('작성 성공');
+      toast.success('작성 성공');
       navigate('/');
     },
     onError: (err) => {
