--- src/hooks/useToast.ts
@@ -1,9 +1,0 @@
-import { toast } from 'react-toastify';
-
-export const useSuccesToast = (message: string) => {
-  toast.success(message);
-};
-
-export const useErrorToast = (message: string) => {
-  toast.error(message);
-};
