+++ src/hooks/useUser.ts
@@ -4,8 +4,8 @@ import { getUserData } from 'apis/user.api';
 import { UserDataType } from 'types/user.type';
 import { ACCESS_KEY } from 'constants/token.constant';
 
-export const GetUserDataFeature = () => {
-  const { data: user, isFetching } = useQuery<UserDataType>(
+export const useUser = () => {
+  const { data: user } = useQuery<UserDataType>(
     [KEY.USER],
     () => getUserData(),
     {
@@ -13,5 +13,11 @@ export const GetUserDataFeature = () => {
     },
   );
 
-  return { user, isFetching };
+  return {
+    user: user || {
+      authority: null,
+      username: null,
+      schoolName: null,
+    },
+  };
 };
